Startupsctipt for Batchdownload
This is the startupscript for the batch mode.
It starts scdl with each track / set / user specified in the file batch.txt.

In the end it adds all downloaded tracks to the 'list'. The list contains the name of every downloaded track. So you can just move the track and the modified version of scdl will not download it again.

Usefull if you have a real big and sorted music-lib.